Change internal function prefix
From double to single underscore. The original intention of the double
underscore was to indicate private/internal functions that are helpers
for other functions and shouldn't be used by/exposed to the end user.

According to the "PEP 8 - Style Guide for Python Code" such intention
should be indicated using a single underscore and double underscore
has another purpose/meaning.

https://www.python.org/dev/peps/pep-0008/#descriptive-naming-styles

Excerpts from the documentation:

> _single_leading_underscore: weak "internal use" indicator.
> E.g. from M import *
> does not import objects whose names start with an underscore.

And:

> __double_leading_underscore:
> when naming a class attribute, invokes name mangling
> (inside class FooBar, __boo becomes _FooBar__boo; see below).