Perform conversion only in one direction
When extracting colors map the color conversion from RGB to CIE L*a*b*
so that the conversion only needs to be performed in one direction.
Conversion between color spaces are prone to slight variations due to
floating precision. Even with the stability corrections done lately this
method works around the "issue" and will a result output that matches
what one would expect.

While restructuring the function I noticed that there were a lot of
manipulation to collections that I suspected cost a bit in memory
management/reallocation. Introducing a flag to store the state and
performing the manipulation in bulk confirmed the suspicion.

As an example: an image with 223715 unique colors showed a dramatic
speed increase, the color extraction went from taking almost 19 minutes
to just above 4 seconds.

    Plus Ultra by Artgerm
    918 x 1300 pixels
    https://www.deviantart.com/artgerm/art/Plus-Ultra-704037787