Don't count fully transparent pixels
Before this change any fully transparent pixel would be counted towards
black which would skew the result. Since the extraction is grouping
colors using CIE L*a*b* there is no easy and natural way to handle
transparency.

On the other hand it feels very natural to remove any pixel that is
fully transparent. It goes well with the more visual grouping tha
CIE L*a*b* allows for which is why it's the color spaced used in this
tool/library.

A long term option would potentially be to introduce options to blend
transparent pixels with a specified color. However there are other
applications that are more suited for such a task.