Support output of colors to GPL
The GIMP color palette is a format create by and for GIMP it's a simple
and open format supported by other applications such as Inkscape and
Krita.

The use case to export directly to a color palette format was provided
by a user of the command line option. A use case I had not though about
but makes perfect sense to support.